+++ examples/array_write_read_ndarray.rs
@@ -1,8 +1,7 @@
 use ndarray::{array, Array2, ArrayD};
 use rayon::iter::{IntoParallelIterator, ParallelIterator};
-use zarrs::{
-    array::storage_transformer::{StorageTransformerExtension, UsageLogStorageTransformer},
-    storage::ReadableWritableListableStorage,
+use zarrs::storage::{
+    storage_adapter::usage_log::UsageLogStorageAdapter, ReadableWritableListableStorage,
 };
 
 fn array_write_read() -> Result<(), Box<dyn std::error::Error>> {
@@ -29,12 +28,9 @@ fn array_write_read() -> Result<(), Box<dyn std::error::Error>> {
                 std::io::stdout(),
                 //    )
             ));
-            let usage_log = Arc::new(UsageLogStorageTransformer::new(log_writer, || {
+            store = Arc::new(UsageLogStorageAdapter::new(store, log_writer, || {
                 chrono::Utc::now().format("[%T%.3f] ").to_string()
             }));
-            store = usage_log
-                .clone()
-                .create_readable_writable_listable_transformer(store);
         }
     }