+++ examples/sharded_array_write_read.rs
@@ -1,8 +1,9 @@
 use itertools::Itertools;
 use zarrs::{
     array::bytes_to_ndarray,
-    array::storage_transformer::{StorageTransformerExtension, UsageLogStorageTransformer},
-    storage::ReadableWritableListableStorage,
+    storage::{
+        storage_adapter::usage_log::UsageLogStorageAdapter, ReadableWritableListableStorage,
+    },
 };
 
 fn sharded_array_write_read() -> Result<(), Box<dyn std::error::Error>> {
@@ -34,12 +35,9 @@ fn sharded_array_write_read() -> Result<(), Box<dyn std::error::Error>> {
                 std::io::stdout(),
                 //    )
             ));
-            let usage_log = Arc::new(UsageLogStorageTransformer::new(log_writer, || {
+            store = Arc::new(UsageLogStorageAdapter::new(store, log_writer, || {
                 chrono::Utc::now().format("[%T%.3f] ").to_string()
             }));
-            store = usage_log
-                .clone()
-                .create_readable_writable_listable_transformer(store);
         }
     }